reproduce_arXiv_1806.01094/requirements.txt
coroICA==0.1.20
h5py>=2.7.1
joblib>=0.11
matplotlib>=2.2.2
mne>=0.16.1
numpy>=1.14.3
pillow>=5.3.0
scipy>=1.1.0
seaborn>=0.8.1
tqdm>=4.23.4
